add payment and payment method modes
add ugly way to handle error responses when creating payment methods
