+++ templates/advantage/_stripe-modal.html
@@ -1,22 +1,21 @@
-<div class="p-modal u-hide" id="renewal-modal" style="overflow: none; position: fixed; z-index: 100;">
+<div
+  class="p-modal u-hide"
+  id="renewal-modal"
+  style="overflow: none; position: fixed; z-index: 100;"
+>
   <div
     class="p-modal__dialog"
     role="dialog"
     aria-labelledby="modal-title"
     aria-describedby="modal-description"
-    style="bottom: 1.5rem; left: 1.5rem; margin: auto; overflow: scroll; position: absolute; right: 1.5rem; top: 1.5rem; width: 600px;"
+    style="width: 700px;"
   >
     <header class="p-modal__header">
-      <h2 class="p-modal__title js-renewal-name" id="modal-title">Renew subscription</h2>
-      <button
-        class="p-modal__close"
-        aria-label="Close active modal"
-        aria-controls="renewal-modal"
-      >
-        Close
-      </button>
+      <h2 class="p-modal__title js-renewal-name" id="modal-title">
+        Renew subscription
+      </h2>
     </header>
-    <p id="modal-description">
+    <div id="modal-description" style="max-height: 70vh; overflow-x: scroll;">
       <ul class="p-list">
         <li class="p-list__item js-renewal-name">Plan type:</li>
         <li class="p-list__item js-renewal-start">Effective from:</li>
@@ -28,38 +27,52 @@
       <form id="payment-form" method="POST" action="/advantage/payment-method">
         <fieldset>
           <label for="name">Name</label>
-          <input name="name" type="text" required>
+          <input
+            name="name"
+            type="text"
+            required
+            value="{{ openid.fullname }}"
+          />
 
           <label for="email">Email</label>
-          <input name="email" type="email" required>
+          <input
+            name="email"
+            type="email"
+            required
+            value="{{ openid.email }}"
+          />
 
           <label for="address">Address</label>
-          <input name="address" type="text">
+          <input name="address" type="text" />
 
           <label for="city">City</label>
-          <input name="city" type="text">
+          <input name="city" type="text" />
 
           <label for="postal_code">ZIP</label>
-          <input name="postal_code" type="text">
+          <input name="postal_code" type="text" />
 
-          {% with raw="true" %}
-          {% include "shared/forms/_country.html" %}
-          {% endwith %}
+          {% with raw="true" %} {% include "shared/forms/_country.html" %} {%
+          endwith %}
 
           <label>Payment information:</label>
 
           <div id="card-errors" class="u-hide"></div>
-          <div id="card-element" class="field" style="background: #fff; border: 1px solid #999; border-radius: 0.125rem; box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.12); padding: 9px;"></div>
+          <div
+            id="card-element"
+            style="
+              background: #fff;
+              border: 1px solid #999;
+              border-radius: 0.125rem;
+              box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.12);
+              padding: 9px;
+            "
+          ></div>
         </fieldset>
-
-        <button class="js-payment-method p-button--action" disabled type="submit">
-          <span class="p-button__copy">Add payment method</span>
-          <i class="p-icon--spinner u-animation--spin"></i>
-        </button>
       </form>
 
       <div id="payment-method-details" class="u-hide">
-        <h3>Payment method:</h3>
+        <h3 class="p-heading--four">Payment method:</h3>
+        <button class="p-button js-reset-modal">Change</button>
         <div class="p-card">
           <div class="row">
             <div class="col-2 js-customer-card-brand"></div>
@@ -71,17 +84,49 @@
           </div>
         </div>
 
-        <p>A receipt will be emailed to:<br /><span class="js-customer-email"></span></p>
-
-        <button class="js-process-payment p-button--action p-button--positive" disabled type="submit">
-          <span class="p-button__copy">Pay</span>
-          <i class="p-icon--spinner is-light u-animation--spin"></i>
-        </button>
+        <p>
+          A receipt will be emailed to:<br /><span
+            class="js-customer-email"
+          ></span>
+        </p>
       </div>
-    </p>
+    </div>
+    <footer class="p-modal__footer u-align--right" style="position: relative;">
+      <i
+        class="p-icon--spinner u-animation--spin u-hide"
+        style="bottom: 0; left: 0; position: absolute;"
+      ></i>
+
+      <button
+        class="js-cancel-modal u-no-margin--bottom"
+        aria-controls="renewal-modal"
+      >
+        Cancel
+      </button>
+
+      <button
+        class="js-payment-method p-button--action u-no-margin--bottom"
+        disabled
+        type="submit"
+      >
+        Save payment details
+      </button>
+
+      <button
+        class="js-process-payment p-button--action p-button--positive u-no-margin--bottom u-hide"
+        disabled
+        type="submit"
+      >
+        Pay
+      </button>
+    </footer>
   </div>
 </div>
 
 <script src="https://js.stripe.com/v3/"></script>
 
-<script src="{{ versioned_static('js/src/stripe.js') }}" defer></script>
+<script
+  src="{{ versioned_static('js/src/renewal-modal.js') }}"
+  type="module"
+  defer
+></script>
